[#763] Fixing possibly flaky tests (only fails on precision):
Note, these might not be flaky long-term, they might only be different
now because of the new algorithm code in a recent commit. Since this
commit message will only ever show up in a squashed commit, I mean
that a commit or two ago, I went back to Greg's algorithm initially
posted in the github comments for this message, which I temporarily
took out and tried a different (incorrect) algorithm with.