fixed frontend and backend tests
This PR includes a major change to the design of the core Group Income protocol in preparation for full accounts support and end-to-end encryption.

The core contract class `GIMessage` has been modified to add the concept of operations. The following operations are supported:

- `OP_CONTRACT` — Creates a new contract chain
- `OP_ACTION` — Performs an end-to-end encrypted action on a contract (e2e cryptography not yet implemented)
- `OP_PUB_ACTION` — Performs a publicly-readable action on a contract
- `OP_KEY_AUTH` — Authorizes a new key to be able to write to a contract
- `OP_KEY_DEAUTH` — Invalidates a key (not implemented)
- `OP_PROTOCOL_UPGRADE` — Advertises a new version of the protocol (not implemented)
- `OP_PROP_SET` — Sets a publicly readable key-value pair on the contract
- `OP_PROP_DEL` — Removes a publicly readable key-value pair from the contract

One of the significant changes that has resulted from this is that the very first message to a contract is no longer the contract's constructor action. Instead, the constructor action is posted second in the chain. The first message merely advertises the existence of a new contract via `OP_CONTRACT`, and the constructor comes next via `OP_ACTION` or `OP_PUB_ACTION`.

- Removed `babel-plugin-transform-inline-environment-variables` and replaced it with `@rollup/plugin-replace`
- Re-enabled the babel cache
- Build process now has two entry points and two outputs: one for the main bundle, and another for a service worker (which is currently unused in the app). As part of this, our build process now correctly supports multiple inputs/outputs via rollup.
- Ran `npm audit fix` to update some deep dependencies